latest changes;

+++ bin/dispose_data.py
@@ -86,16 +86,17 @@ def _main_func(description):
                 shutil.rmtree(os.path.join(dout_s_root,_dir))
         atmhistpath = os.path.join(dout_s_root,"atm","hist")
         icehistpath = os.path.join(dout_s_root,"ice","hist")
-        for histfile in os.listdir(atmhistpath):
-            if "h1" in histfile or "h4" in histfile:
-                os.unlink(os.path.join(atmhistpath,histfile))
+        if os.path.isdir(atmhistpath):
+            for histfile in os.listdir(atmhistpath):
+                if "h1" in histfile or "h4" in histfile:
+                    os.unlink(os.path.join(atmhistpath,histfile))
         #Concatinate cice history into a single file
-
-        fnameout = basecasename+"."+basemonth+"."+date+"."+member+".cice.h.nc"
-        run_cmd("ncrcat * "+fnameout,from_dir=icehistpath)
-        for _file in glob.iglob(os.path.join(icehistpath,"*ice.h.*.nc")):
-            os.unlink(_file)
-        send_data_to_campaignstore(dout_s_root+os.sep )
+        if os.path.isdir(icehistpath):
+            fnameout = basecasename+"."+basemonth+"."+date+"."+member+".cice.h.nc"
+            run_cmd("ncrcat * "+fnameout,from_dir=icehistpath)
+            for _file in glob.iglob(os.path.join(icehistpath,"*ice.h.*.nc")):
+                os.unlink(_file)
+            send_data_to_campaignstore(dout_s_root+os.sep )
         
         
 if __name__ == "__main__":
